Separate the styling code to separate files with name style.js Relates #60
