+++ client/src/Components/LinkItem/style.js
@@ -1,0 +1,18 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles(() => ({
+  root: {
+    marginLeft: '1.25rem',
+    fontWeight: 'bold',
+    color: '#2F5C84',
+  },
+  signupLink: {
+    color: '#FFFFFF !important',
+    backgroundColor: '#EB9235',
+    padding: '1rem',
+    textAlign: 'center',
+    borderRadius: '3px',
+  },
+}));
+
+export default useStyles;
